+++ src/marker/Generator.java
@@ -5,8 +5,6 @@ import java.util.List;
 import java.util.Random;
 import java.util.stream.IntStream;
 
-import static model.kripke.KripkeGenerator.rand;
-
 public interface Generator {
     Random rand = new Random();
